+++ grunt/options/karma.js
@@ -31,20 +31,5 @@ module.exports = {
       // Use Phantom and Firefox for Travis
       browsers: ['PhantomJS', 'Firefox']
     }
-  },
-
-  // Production-ready JS
-  build: {
-    options: {
-
-      // Use Phantom for Jenkins
-      browsers: ['PhantomJS'],
-
-      files: [
-        'dist/<%= pkg.version %>/main.min.js',
-        'test/dist/tests.min.js',
-        'test/test-main.js'
-      ]
-    }
   }
 };